backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
httpx>=0.27.0
PyJWT==2.8.0
cryptography==41.0.7
python-dotenv==1.0.0
